+++ frontend/src/components/__tests__/AnimatedBackground.test.tsx
