+++ frontend/src/components/__tests__/Board.test.tsx
