+++ frontend/src/components/__tests__/CaptchaType67.test.tsx
