+++ frontend/src/components/__tests__/ContentArea.test.tsx
