+++ frontend/src/components/__tests__/FloatingMenu.test.tsx
