+++ frontend/src/components/__tests__/Sidebar.test.tsx
