+++ frontend/src/components/__tests__/createNewSpace.test.tsx
