+++ frontend/src/components/__tests__/passwordValidator.test.ts
